+++ src/zimscraperlib/zim/metadata.py
@@ -29,7 +29,10 @@ def validate_required_values(name: str, value: Any):
         raise ValueError(f"Missing value for {name}")
 
 
-def validate_standard_str_types(name: str, value: str):
+def validate_standard_str_types(
+    name: str,
+    value: str | bytes,
+):
     """ensures standard string metadata are indeed str"""
     if name in (
         "Name",
@@ -45,7 +48,7 @@ def validate_standard_str_types(name: str, value: str):
         "Source",
         "Scraper",
     ) and not isinstance(value, str):
-        raise ValueError(f"Invalid type for {name}")
+        raise ValueError(f"Invalid type for {name}: {type(value)}")
 
 
 def validate_title(name: str, value: str):
@@ -58,16 +61,13 @@ def validate_date(name: str, value: datetime.datetime | datetime.date | str):
     """ensures Date metadata can be casted to an ISO 8601 string"""
     if name == "Date":
         if not isinstance(value, (datetime.datetime, datetime.date, str)):
-            raise ValueError(f"Invalid type for {name}.")
+            raise ValueError(f"Invalid type for {name}: {type(value)}")
         elif isinstance(value, str):
             match = re.match(r"(?P<year>\d{4})-(?P<month>\d{2})-(?P<day>\d{2})", value)
+            if not match:
+                raise ValueError(f"Invalid {name} format, not matching regex")
             try:
-                datetime.date(
-                    **{
-                        k: int(v)
-                        for k, v in match.groupdict().items()  # pyright: ignore
-                    }
-                )
+                datetime.date(**{k: int(v) for k, v in match.groupdict().items()})
             except Exception as exc:
                 raise ValueError(f"Invalid {name} format: {exc}") from None
 
@@ -82,7 +82,7 @@ def validate_language(name: str, value: Iterable[str] | str):
                 raise ValueError(f"{code} is not ISO-639-3.")
 
 
-def validate_counter(name: str, value: str):  # noqa: ARG001
+def validate_counter(name: str, _: str):
     """ensures Counter metadata is not manually set"""
     if name == "Counter":
         raise ValueError(f"{name} cannot be set. libzim sets it.")